refactor(project): rename registry by lock
