+++ src/Project/InitializedProject.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Haspadar\Piqule\Project;
 
 use Haspadar\Piqule\PiquleException;
-use Haspadar\Piqule\Project\Registry\Registry;
+use Haspadar\Piqule\Project\Lock\Lock;
 use Haspadar\Piqule\Source\SourceDirectory;
 use Haspadar\Piqule\Target\DiskTarget;
 use Haspadar\Piqule\Target\Materialization\Materialization;
@@ -18,12 +18,12 @@ final readonly class InitializedProject implements Project
         private TargetDirectory $targetDirectory,
     ) {}
 
-    public function init(Materialization $materialization, Registry $lock): void
+    public function init(Materialization $materialization, Lock $lock): void
     {
         throw new PiquleException('Project is already initialized');
     }
 
-    public function update(Materialization $materialization, Registry $lock): void
+    public function update(Materialization $materialization, Lock $lock): void
     {
         foreach ($this->sourceDirectory->files() as $sourceFile) {
             $lock = $materialization->applyTo(
